+++ internal/dataplane/sendconfig/inmemory.go
@@ -51,20 +51,21 @@ func NewUpdateStrategyInMemory(
 func (s UpdateStrategyInMemory) Update(ctx context.Context, targetState ContentWithHash) (
 	err error,
 	resourceErrors []ResourceError,
+	rawErrBody []byte,
 	resourceErrorsParseErr error,
 ) {
 	dblessConfig := s.configConverter.Convert(targetState.Content)
 	config, err := json.Marshal(dblessConfig)
 	if err != nil {
-		return fmt.Errorf("constructing kong configuration: %w", err), nil, nil
+		return fmt.Errorf("constructing kong configuration: %w", err), nil, nil, nil
 	}
 
 	if errBody, err := s.configService.ReloadDeclarativeRawConfig(ctx, bytes.NewReader(config), true, true); err != nil {
 		resourceErrors, parseErr := parseFlatEntityErrors(errBody, s.logger)
-		return err, resourceErrors, parseErr
+		return err, resourceErrors, errBody, parseErr
 	}
 
-	return nil, nil, nil
+	return nil, nil, nil, nil
 }
 
 func (s UpdateStrategyInMemory) MetricsProtocol() metrics.Protocol {