+++ test/kongintegration/translator_golden_tests_outputs_test.go
@@ -127,7 +127,7 @@ func ensureGoldenTestOutputIsAccepted(
 	require.NoError(t, err)
 
 	require.Eventually(t, func() bool {
-		err, resourceErrors, parseErr := sut.Update(ctx, sendconfig.ContentWithHash{Content: content})
+		err, resourceErrors, rawErrBody, parseErr := sut.Update(ctx, sendconfig.ContentWithHash{Content: content})
 		if err != nil {
 			t.Logf("error: %v", err)
 			return false
@@ -138,6 +138,7 @@ func ensureGoldenTestOutputIsAccepted(
 		}
 		if parseErr != nil {
 			t.Logf("parse error: %v", parseErr)
+			t.Logf("raw error: %v", rawErrBody)
 			return false
 		}
 		return true